Use the filter pills in the InventoryV2/Search component within the 'inventory for species|nursery' views
