+++ src/components/InventoryV2/Search.tsx
@@ -1,6 +1,7 @@
-import React, { useEffect, useMemo, useState } from 'react';
-import { Box, Popover, useTheme } from '@mui/material';
-import { Button, Textfield, Tooltip } from '@terraware/web-components';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { Grid, Box, Popover, useTheme } from '@mui/material';
+import { Button, PillListItem, Textfield, Tooltip } from '@terraware/web-components';
+import { PillList } from '@terraware/web-components';
 import { makeStyles } from '@mui/styles';
 import strings from 'src/strings';
 import { useLocalization, useOrganization } from 'src/providers/hooks';
@@ -18,7 +19,7 @@ import { selectProjects } from 'src/redux/features/projects/projectsSelectors';
 import { Project } from 'src/types/Project';
 import InventoryFilters, { InventoryFiltersType } from 'src/components/InventoryV2/InventoryFilter';
 import { OriginPage } from 'src/components/InventoryV2/InventoryBatch';
-import { convertFilterGroupToMap } from 'src/components/InventoryV2/FilterUtils';
+import { convertFilterGroupToMap, getNurseryName } from 'src/components/InventoryV2/FilterUtils';
 import isEnabled from 'src/features';
 
 const useStyles = makeStyles(() => ({
@@ -49,6 +50,11 @@ interface SearchProps {
   showProjectsFilter?: boolean;
 }
 
+type PillListItemWithEmptyValue = Omit<PillListItem<string>, 'id'> & {
+  id: keyof InventoryFiltersType;
+  emptyValue: unknown;
+};
+
 export default function Search(props: SearchProps): JSX.Element | null {
   const { searchValue, onSearch, filters, setFilters, showProjectsFilter } = props;
 
@@ -71,6 +77,8 @@ export default function Search(props: SearchProps): JSX.Element | null {
 
   const subLocations = useAppSelector(selectSubLocations);
 
+  const [filterPillData, setFilterPillData] = useState<PillListItemWithEmptyValue[]>([]);
+
   const [filterAnchorEl, setFilterAnchorEl] = useState<null | HTMLElement>(null);
   const handleFilterClick = (event: React.MouseEvent<HTMLElement>) => setFilterAnchorEl(event.currentTarget);
   const handleFilterClose = () => setFilterAnchorEl(null);
@@ -90,11 +98,102 @@ export default function Search(props: SearchProps): JSX.Element | null {
     [activeLocale]
   );
 
+  const getSpeciesName = useCallback(
+    (speciesId: number) => (species || []).find((s) => s.id === speciesId)?.scientificName,
+    [species]
+  );
+
+  const getSubLocationName = useCallback(
+    (subLocationId: number) => subLocations?.find((sl) => subLocationId === sl.id)?.name ?? '',
+    [subLocations]
+  );
+
+  const getProjectName = useCallback(
+    (projectId: number) => (projects || []).find((p) => p.id === projectId)?.name,
+    [projects]
+  );
+
+  useEffect(() => {
+    const data: PillListItemWithEmptyValue[] = [];
+    if (filters.facilityIds?.length) {
+      data.push({
+        id: 'facilityIds',
+        label: strings.NURSERY,
+        value: filters.facilityIds?.map((id) => getNurseryName(id, selectedOrganization)).join(', ') ?? '',
+        emptyValue: [],
+      });
+    }
+
+    if (filters.speciesIds?.length) {
+      data.push({
+        id: 'speciesIds',
+        label: strings.SPECIES,
+        value: filters.speciesIds?.map(getSpeciesName).join(', ') ?? '',
+        emptyValue: [],
+      });
+    }
+
+    if (filters.subLocationsIds?.length) {
+      data.push({
+        id: 'subLocationsIds',
+        label: strings.SUB_LOCATIONS,
+        value: filters.subLocationsIds?.map(getSubLocationName).join(', ') ?? '',
+        emptyValue: [],
+      });
+    }
+
+    if (filters.projectIds?.length) {
+      data.push({
+        id: 'projectIds',
+        label: strings.PROJECTS,
+        value: filters.projectIds?.map(getProjectName).join(', ') ?? '',
+        emptyValue: [],
+      });
+    }
+
+    if (filters.showEmptyBatches && filters.showEmptyBatches[0] === 'true') {
+      data.push({
+        id: 'showEmptyBatches',
+        value: strings.FILTER_SHOW_EMPTY_BATCHES,
+        emptyValue: ['false'],
+      });
+    }
+
+    setFilterPillData(data);
+  }, [
+    selectedOrganization,
+    filters.facilityIds,
+    filters.speciesIds,
+    filters.subLocationsIds,
+    filters.projectIds,
+    filters.showEmptyBatches,
+    getSpeciesName,
+    getSubLocationName,
+    getProjectName,
+  ]);
+
   useEffect(() => {
     if (origin === 'Nursery') {
       void dispatch(requestSpecies(selectedOrganization.id));
     }
   }, [origin, dispatch, selectedOrganization.id]);
+
+  const onRemovePillList = useCallback(
+    (filterId: keyof InventoryFiltersType) => {
+      const filter = filterPillData?.find((filterPillDatum) => filterPillDatum.id === filterId);
+      if (filterId === 'facilityIds') {
+        setFilters({ ...filters, facilityIds: [], subLocationsIds: [] });
+      } else if (filterId === 'showEmptyBatches') {
+        setFilterGroupFilters({
+          showEmptyBatches: { ...initialFilters.showEmptyBatches, values: ['false'] },
+        });
+        setFilters({ ...filters, showEmptyBatches: ['false'] });
+      } else {
+        setFilters({ ...filters, [filterId]: filter?.emptyValue || null });
+      }
+    },
+    [filterPillData, filters, setFilters, setFilterGroupFilters]
+  );
 
   if (origin === 'Nursery' && !species && !projects) {
     return null;
@@ -173,7 +272,7 @@ export default function Search(props: SearchProps): JSX.Element | null {
           />
         )}
 
-        {origin === 'Batches' && (
+        {['Batches', 'Species', 'Nursery'].includes(origin) && (
           <Box sx={{ marginTop: theme.spacing(0.5) }}>
             <Tooltip title={strings.FILTER}>
               <Button
@@ -217,6 +316,9 @@ export default function Search(props: SearchProps): JSX.Element | null {
           </Box>
         )}
       </Box>
+      <Grid display='flex' flexDirection='row' alignItems='center' sx={{ marginTop: theme.spacing(2) }}>
+        <PillList data={filterPillData} onRemove={onRemovePillList} />
+      </Grid>
     </>
   );
 }
