+++ src/components/InventoryV2/view/InventorySeedlingsTable.tsx
@@ -1,8 +1,8 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import { Typography, Grid, Box, useTheme, Theme, Popover } from '@mui/material';
+import { Typography, Grid, Box, useTheme, Theme } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import { Button, DropdownItem, PillList, PillListItem, TableColumnType, Tooltip } from '@terraware/web-components';
+import { Button, DropdownItem, TableColumnType } from '@terraware/web-components';
 import { TopBarButton } from '@terraware/web-components/components/table';
 import strings from 'src/strings';
 import useDebounce from 'src/utils/useDebounce';
@@ -13,15 +13,10 @@ import { NurseryBatchService } from 'src/services';
 import useSnackbar from 'src/utils/useSnackbar';
 import { APP_PATHS } from 'src/constants';
 import { useLocalization, useOrganization } from 'src/providers';
-import { useAppSelector } from 'src/redux/store';
-import { selectSpecies } from 'src/redux/features/species/speciesSelectors';
-import { selectSubLocations } from 'src/redux/features/subLocations/subLocationsSelectors';
-import { selectProjects } from 'src/redux/features/projects/projectsSelectors';
 import Table from 'src/components/common/table';
 import { SortOrder } from 'src/components/common/table/sort';
 import OptionsMenu from 'src/components/common/OptionsMenu';
-import FilterGroup, { FilterField } from 'src/components/common/FilterGroup';
-import { convertFilterGroupToMap, getNurseryName, isBatchEmpty } from 'src/components/InventoryV2/FilterUtils';
+import { isBatchEmpty } from 'src/components/InventoryV2/FilterUtils';
 import { InventoryFiltersType } from 'src/components/InventoryV2/InventoryFilter';
 import Search from 'src/components/InventoryV2/Search';
 import BatchesCellRenderer from './BatchesCellRenderer';
@@ -56,30 +51,6 @@ export interface InventorySeedlingsTableProps {
   ) => Promise<SearchResponseElement[] | null>;
 }
 
-const useStyles = makeStyles((theme: Theme) => ({
-  popoverContainer: {
-    '& .MuiPaper-root': {
-      border: `1px solid ${theme.palette.TwClrBrdrTertiary}`,
-      borderRadius: '8px',
-      width: '480px',
-    },
-  },
-}));
-
-const initialFilters: Record<string, SearchNodePayload> = {
-  showEmptyBatches: {
-    field: 'showEmptyBatches',
-    values: ['false'],
-    type: 'Exact',
-    operation: 'field',
-  },
-};
-
-type PillListItemWithEmptyValue = Omit<PillListItem<string>, 'id'> & {
-  id: keyof InventoryFiltersType;
-  emptyValue: unknown;
-};
-
 export default function InventorySeedlingsTable(props: InventorySeedlingsTableProps): JSX.Element {
   const {
     modified,
@@ -98,14 +69,8 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
   const { selectedOrganization } = useOrganization();
   const { isMobile, isDesktop } = useDeviceInfo();
   const theme = useTheme();
-  const classes = useStyles();
   const snackbar = useSnackbar();
   const history = useHistory();
-  const { activeLocale } = useLocalization();
-
-  const species = useAppSelector(selectSpecies);
-  const projects = useAppSelector(selectProjects);
-  const subLocations = useAppSelector(selectSubLocations);
 
   const [openExportModal, setOpenExportModal] = useState<boolean>(false);
   const [temporalSearchValue, setTemporalSearchValue] = useState<string>('');
@@ -114,13 +79,11 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
   // the filtered-by-search batches list that comes back from the API
   const [filteredBatches, setFilteredBatches] = useState<any[]>([]);
   const [filters, setFilters] = useForm<InventoryFiltersType>({});
-  const [filterGroupFilters, setFilterGroupFilters] = useForm<Record<string, SearchNodePayload>>(initialFilters);
   const [selectedRows, setSelectedRows] = useState<any[]>([]);
   const [openDeleteModal, setOpenDeleteModal] = useState<boolean>(false);
   const [openNewBatchModal, setOpenNewBatchModal] = useState<boolean>(false);
   const [selectedBatch, setSelectedBatch] = useState<any>();
   const [searchSortOrder, setSearchSortOrder] = useState<SearchSortOrder>();
-  const [filterPillData, setFilterPillData] = useState<PillListItemWithEmptyValue[]>([]);
 
   const debouncedSearchTerm = useDebounce(temporalSearchValue, 250);
 
@@ -297,25 +260,6 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
     }
   };
 
-  const [filterAnchorEl, setFilterAnchorEl] = useState<null | HTMLElement>(null);
-  const handleFilterClick = (event: React.MouseEvent<HTMLElement>) => setFilterAnchorEl(event.currentTarget);
-  const handleFilterClose = () => setFilterAnchorEl(null);
-
-  const filterGroupColumns = useMemo<FilterField[]>(
-    () =>
-      activeLocale
-        ? [
-            {
-              name: 'showEmptyBatches',
-              label: strings.FILTER_SHOW_EMPTY_BATCHES,
-              showLabel: false,
-              type: 'boolean',
-            },
-          ]
-        : [],
-    [activeLocale]
-  );
-
   const batchesExport = useCallback(() => {
     if (!originId || !getBatchesExport) {
       return Promise.resolve([] as SearchResponseElement[]);
@@ -323,97 +267,6 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
 
     return getBatchesExport(selectedOrganization.id, originId, getSearchFields(), searchSortOrder);
   }, [getBatchesExport, selectedOrganization.id, originId, getSearchFields, searchSortOrder]);
-
-  const onRemovePillList = useCallback(
-    (filterId: keyof InventoryFiltersType) => {
-      const filter = filterPillData?.find((filterPillDatum) => filterPillDatum.id === filterId);
-      if (filterId === 'facilityIds') {
-        setFilters({ ...filters, facilityIds: [], subLocationsIds: [] });
-      } else if (filterId === 'showEmptyBatches') {
-        setFilterGroupFilters({
-          showEmptyBatches: { ...initialFilters.showEmptyBatches, values: ['false'] },
-        });
-        setFilters({ ...filters, showEmptyBatches: ['false'] });
-      } else {
-        setFilters({ ...filters, [filterId]: filter?.emptyValue || null });
-      }
-    },
-    [filterPillData, filters, setFilters, setFilterGroupFilters]
-  );
-
-  const getSpeciesName = useCallback(
-    (speciesId: number) => (species || []).find((s) => s.id === speciesId)?.scientificName,
-    [species]
-  );
-
-  const getSubLocationName = useCallback(
-    (subLocationId: number) => subLocations?.find((sl) => subLocationId === sl.id)?.name ?? '',
-    [subLocations]
-  );
-
-  const getProjectName = useCallback(
-    (projectId: number) => (projects || []).find((p) => p.id === projectId)?.name,
-    [projects]
-  );
-
-  useEffect(() => {
-    const data: PillListItemWithEmptyValue[] = [];
-    if (filters.facilityIds?.length) {
-      data.push({
-        id: 'facilityIds',
-        label: strings.NURSERY,
-        value: filters.facilityIds?.map((id) => getNurseryName(id, selectedOrganization)).join(', ') ?? '',
-        emptyValue: [],
-      });
-    }
-
-    if (filters.speciesIds?.length) {
-      data.push({
-        id: 'speciesIds',
-        label: strings.SPECIES,
-        value: filters.speciesIds?.map(getSpeciesName).join(', ') ?? '',
-        emptyValue: [],
-      });
-    }
-
-    if (filters.subLocationsIds?.length) {
-      data.push({
-        id: 'subLocationsIds',
-        label: strings.SUB_LOCATIONS,
-        value: filters.subLocationsIds?.map(getSubLocationName).join(', ') ?? '',
-        emptyValue: [],
-      });
-    }
-
-    if (filters.projectIds?.length) {
-      data.push({
-        id: 'projectIds',
-        label: strings.PROJECTS,
-        value: filters.projectIds?.map(getProjectName).join(', ') ?? '',
-        emptyValue: [],
-      });
-    }
-
-    if (filters.showEmptyBatches && filters.showEmptyBatches[0] === 'true') {
-      data.push({
-        id: 'showEmptyBatches',
-        value: strings.FILTER_SHOW_EMPTY_BATCHES,
-        emptyValue: ['false'],
-      });
-    }
-
-    setFilterPillData(data);
-  }, [
-    selectedOrganization,
-    filters.facilityIds,
-    filters.speciesIds,
-    filters.subLocationsIds,
-    filters.projectIds,
-    filters.showEmptyBatches,
-    getSpeciesName,
-    getSubLocationName,
-    getProjectName,
-  ]);
 
   return (
     <>
@@ -485,14 +338,7 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
             </Box>
           </Box>
 
-          <Box
-            sx={{
-              display: 'flex',
-              justifyContent: 'flex-start',
-              alignItems: 'center',
-              marginBottom: theme.spacing(2),
-            }}
-          >
+          <Box>
             <Search
               searchValue={temporalSearchValue}
               onSearch={(val) => setTemporalSearchValue(val)}
@@ -501,54 +347,7 @@ export default function InventorySeedlingsTable(props: InventorySeedlingsTablePr
               origin={origin}
               showProjectsFilter
             />
-
-            <Box sx={{ marginTop: theme.spacing(0.5) }}>
-              <Tooltip title={strings.FILTER}>
-                <Button
-                  id='batchFilters'
-                  onClick={(event) => event && handleFilterClick(event)}
-                  type='passive'
-                  priority='ghost'
-                  icon='filter'
-                />
-              </Tooltip>
-              <Popover
-                id='simple-popover'
-                open={Boolean(filterAnchorEl)}
-                anchorEl={filterAnchorEl}
-                onClose={handleFilterClose}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'center',
-                }}
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'center',
-                }}
-                className={classes.popoverContainer}
-              >
-                <FilterGroup
-                  initialFilters={filterGroupFilters}
-                  fields={filterGroupColumns}
-                  onConfirm={(_filterGroupFilters: Record<string, SearchNodePayload>) => {
-                    handleFilterClose();
-                    setFilterGroupFilters(_filterGroupFilters);
-                    setFilters({ ...filters, ...convertFilterGroupToMap(_filterGroupFilters) });
-                  }}
-                  onCancel={handleFilterClose}
-                />
-              </Popover>
-            </Box>
           </Box>
-
-          <Grid
-            display='flex'
-            flexDirection='row'
-            alignItems='center'
-            sx={{ marginTop: theme.spacing(0.5), marginLeft: theme.spacing(1) }}
-          >
-            <PillList data={filterPillData} onRemove={onRemovePillList} />
-          </Grid>
 
           <Box marginTop={theme.spacing(2)}>
             <Table
